+++ test/irmin-pack/common.mli
@@ -19,7 +19,7 @@ end) : sig
 
   type d = { dict : Dict.t; clone : readonly:bool -> Dict.t }
 
-  val get_dict : ?version:Irmin_pack.version -> unit -> d
+  val get_dict : unit -> d
   (** Fresh, empty dict. *)
 
   type t = {
@@ -29,7 +29,7 @@ end) : sig
     clone_index_pack : readonly:bool -> (Index.t * [ `Read ] Pack.t) Lwt.t;
   }
 
-  val get_pack : ?version:Irmin_pack.version -> ?lru_size:int -> unit -> t Lwt.t
+  val get_pack : ?lru_size:int -> unit -> t Lwt.t
   (** Fresh, empty index and pack. [clone_pack] opens a clone of the pack at the
       same location, [clone_index_pack] opens a clone of the index and the pack. *)
 
@@ -41,5 +41,3 @@ val get : 'a option -> 'a
 val sha1 : string -> H.t
 
 val rm_dir : string -> unit
-
-val version : Irmin_pack.version Alcotest.testable