+++ CHANGES.md
@@ -31,6 +31,7 @@
 - **irmin-pack**:
   - Added `index_throttle` option to `Irmin_pack.config`, which exposes the
     memory throttle feature of `Index` in `Irmin-Pack`. (#1049, @icristescu)
+  - Added `Pack.clear` and `Dict.clear` (#1047, @icristescu, @CraigFe, @samoht)
 
 #### Changed
 
@@ -45,6 +46,9 @@
     files on disk. (#1008, @icristescu)
   - Renamed `sync` to `flush` for the operation that flushes to disk all buffers
     of a read-write instance. (#1008, @icristescu)
+  - Changed the format of headers for the files on disk to include a generation
+    number. Version 1 of irmin-pack was used for the previous format, version 2
+    is used with the new format. (#1047, @icristescu, @CraigFe, @samoht)
 
 ### 2.2.0 (2020-06-26)